Safari-611.1.10/LayoutTests/fast/performance/performance-now-timestamps-expected.txt

Tests the timestamps from performance.now increase and are in milliseconds

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS firstTimestamp is defined.
PASS firstTimestamp is >= 0
PASS firstTimestamp < 5000 is true
PASS secondTimestamp is defined.
PASS secondTimestamp is >= firstTimestamp + (waitTime / 2)
PASS elapsed is >= 1
PASS elapsed < 100 is true
PASS successfullyParsed is true

TEST COMPLETE
